Comment out auctex packages
Comments out company-auctex and the laTeX packages in the
use-package list.
At the time of commiting I am having an issue with the build
trying to find an old version of the elpa auctex package
which is no longer provided.

For now removing auctex is not a big deal to me; I haven't used it
and it was only in the list from the original bauer package list.

I would still like to find the root cause of this issue though
and see if it can be fixed. Why doesn't `use-package' try to get
the latest version of the package?