Change emacsGcc to emacsNativeComp
When I ran a build using the freshly updated flake lock file, I was
given a warning that 'emacsGcc' is deprecated and that 'emacsNativeComp'
should be used in its place.

It appears that it allows the use of 'libgccjit' in compiliation, which
has some expected speed improvements.

In anycase, given the deprecation of 'emacsGcc', I thought it best to
make the change to the suggested funciton sooner than later.